@@ -6,23 +6,20 @@ const initialState = {
   isLoading: false,
   user: null,
 };
-export const loginUser = createAsyncThunk(
-  "/auth/login",
-  async (formData) => {
-    const response = await axios.post(
-      "http://localhost:5000/api/auth/login",
-      formData,
-      {
-        withCredentials: true,
-      }
-    );
-    return response.data;
-  }
-);
+export const loginUser = createAsyncThunk("/auth/login", async (formData) => {
+  const response = await axios.post(
+    "http://localhost:5000/api/auth/login",
+    formData,
+    {
+      withCredentials: true,
+    }
+  );
+  return response.data;
+});
 export const registerUser = createAsyncThunk(
   "/auth/register",
   async (formData) => {
-    const response = await axios.post( 
+    const response = await axios.post(
       "http://localhost:5000/api/auth/register",
       formData,
       {
@@ -55,7 +52,8 @@ const authSlice = createSlice({
         state.isLoading = false;
         state.user = null;
         state.isAuthenticated = false;
-      }).addCase(loginUser.pending, (state) => {
+      })
+      .addCase(loginUser.pending, (state) => {
         state.isLoading = true;
       })
       .addCase(loginUser.fulfilled, (state, action) => {
